dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
anthropic==0.40.0
sse-starlette==2.1.3
